@@ -1,12 +1,6 @@
 require("dotenv").config();
 module.exports = {
   initmail() {
-    /*  let initMailTemplate = fs.readFile(__dirname + "/g.html", "utf8", (data) => {
-        return data
-    })
-
-    console.log("this is mail", initMailTemplate);
- */
     return `<!DOCTYPE html>
     <html lang="en">
       <head>
@@ -34,8 +28,10 @@ module.exports = {
       <body>
         <h1>Please validate your E-Mail address to use the drunc service</h1>
         <p>just click this Button and the magic Wizard will validate your account</p>
-        <a href="${process.env.NODE_ENV == 'development' ? 'http://localhost:8080' : 'drunc.net'}/#/varifyer?verificationID=${id}">Validate me</a>
-      </body>
+        <a href="${process.env.NODE_ENV == 'development' ? 'http://localhost:8080' : 'drunc.net'}/varifyer?verificationID=${id}">Validate me</a>
+        <p>Your mailer dosent support HTML ??? try copy this link in the browsers searchbar</p>
+        <p>${process.env.NODE_ENV == 'development' ? 'http://localhost:8080' : 'drunc.net'}/varifyer?verificationID=${id}</p>
+        </body>
     </html>
     `;
   },
@@ -51,7 +47,7 @@ module.exports = {
       <body>
         <h1>Someone has invited you to his project</h1>
         <p>checkout your account to see stuff</p>
-        <a href="${process.env.NODE_ENV == "development" ? "http://localhost:8080" : "drunc.net"}/#/home">go To your account</a>
+        <a href="${process.env.NODE_ENV == "development" ? "http://localhost:8080" : "drunc.net"}/home">go To your account</a>
       </body>
     </html>
     `;
