@@ -2,8 +2,8 @@
   <div class="projectCluster_wrapper">
     <p v-if="projects.length == 0">Looks quiet empty here</p>
     <div class="projectcluster_preview">
-      <ProjectPreview class="projectcluster_project" :project="{ id: 'addProject' }" @click="goToProject('-1')" v-if="hasAdd" />
-      <div class="projectcluster_project" v-for="proj in projects" :key="proj" @click="goToProject(proj)">
+      <ProjectPreview class="projectcluster_project" :project="{ id: 'addProject' }" @mouseup="goToProject('-1')" v-if="hasAdd" />
+      <div class="projectcluster_project" v-for="proj in projects" :key="proj" @mouseup="goToProject(proj)">
         <ProjectPreview :project="{ id: proj }" />
         <div v-if="sharedbyself"></div>
         <p>{{ proj }}</p>
@@ -28,7 +28,13 @@ export default {
     sharedfrom: { type: Boolean, default: false },
   },
   methods: {
-    async goToProject(projectID) {
+    async goToProject(goto) {
+      let projectID;
+      if (typeof goto == "object") {
+        projectID = goto.projectID;
+      } else {
+        projectID = goto;
+      }
       const id = await api.projectcall("projects/getID", {
         SID: localStorage.getItem("sessionID"),
         user: localStorage.getItem("usr"),
